Refactor clean arch export report
